requirements.txt
# Telegram-бот 
python-telegram-bot==21.2

# Загрузка / генерация 
torch>=2.2         
transformers>=4.40
diffusers>=0.28
accelerate>=0.29
torchvision==0.18.1
diffusers>=0.28

# Обработка изображений 
Pillow>=10.0
opencv-python       

# Прочее
tqdm
hf_xet
huggingface_hub>=0.16.4
deep-translator
